Moved misc contents to utils
